Moved CountingMemoryResource to separate file (under Aql).
